requirements.txt
langgraph>=0.4.0
litellm>=1.60.0
sqlglot>=26.0.0
sqlalchemy>=2.0.0
numpy>=1.26.0
pandas>=2.2.0
pandera>=0.22.0
scikit-learn>=1.4.0
fastapi>=0.115.0
uvicorn>=0.34.0
slowapi>=0.1.9
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.0
croniter>=1.4.0
pytest-asyncio>=0.23.0
